Centralize the theme interface
